[draft for NHWC/NCHW stride fix]
Changing the way we compute channels_last stride. Use stride for size 1 dimension to implicitly carry intended memory
layout.

Note that:
This gives improvements the flexibility for user to express intended layout on more cases, but it DOES NOT solve the
fundamental issue from our current proposal [Ambiguity in our stride representation].

for example:
1. Tensor with size N111
2. Expanded tensor

Our current WAR is to always fallback to NCHW in ambiguous cases.